requirements.txt
Telethon==1.21.1
cryptg==0.2.post2
Pillow==8.2.0
PyYAML==5.3
